test: extract shared PBT arbs to PbtArbs
